Debian package: Add evms-cli to Suggests.
EVMS has been removed[1] from Debian before Lenny, but it is necessary
for some optional functionality of xen-tools. And since some Debian
derived distributions (e.g. Ubuntu 6.06 LTS Server, Ubuntu 8.04 LTS
and grml) still support it, it's included in the list of suggested
dependencies for the sake of completeness and correctness.

  [1] http://bugs.debian.org/491808