Revert "Generate ChangeLog during package build"
This reverts commit 0cb518bc5af166b9a221bc5780f323a563158779 and fixes
http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=595883

The changelog MUST not be generated during the Debian package build as
it relies on a checked out copy of the VCS repository -- what the
Debian package isn't. The Debian (and any other packaging) gets the
changelog from our tar.gz balls.

We will have to fix the issue, which the reverted commit tried to fix,
differently.