No more use $#array in boolean context
This fixes a regression introduced in
503db1668b595d71b45a94d7391d5c9e6b9e9d1e.

`@array = undef` doesn't empty an array but fills it with one item
which is undef. Hence `$#array` is zero and hence false.

Since the commit above, `@array = ()` was used instead of `@array =
undef`, hence `$#array` no more was false when the array seemed to be
empty (but wasn't). This caused an empty partition table in the
configuration file and hence prohibited the start of the DomU.

This commit fixes the according `!$#array` to just `!@array` and adds
the according changelog entry for both commits. (I initially thought
that such a tiny commit like the one above is not worth being
mentioned. I obviously was wrong.)