Depend and Build-Depend on non-(build-)essential package mount
Closes: #1027383

Thanks to Santiago Vila for the bug report.

Actually the build-dependency is not strictly necessary as mount is
not really used at build or test time. But there's one test checking
the environment for a typical make, make test, make install workflow
to abort if not all required run-time dependencies can be found.

Then again, for the same reason as given in the bug report, a run-time
dependency on mount was missing, too.